Enhance sandbox deletion wait logic and remove obsolete proxy tests
- Updated the `wait_for_sandbox_deletion` function to check for both non-existence and `TERMINATED` status of the sandbox, improving the accuracy of deletion checks.
- Removed the `test_proxy.py` file, which contained outdated tests related to proxy configurations, streamlining the test suite.
- Introduced new helper files and tests for proxy integration, ensuring comprehensive coverage of proxy functionalities in the sandbox environment.
